refactor(dao): Extract a function to query issues
Move functionality to query for issues from `IssueService` to
`IssuesDao`. This makes it possible to query for the issues of a scan
summary as well.
