+++ dao/src/main/kotlin/tables/shared/IssuesTable.kt
@@ -30,6 +30,8 @@ import org.eclipse.apoapsis.ortserver.model.runs.Issue
 
 import org.jetbrains.exposed.dao.LongEntity
 import org.jetbrains.exposed.dao.id.EntityID
+import org.jetbrains.exposed.sql.Query
+import org.jetbrains.exposed.sql.ResultRow
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.stringLiteral
 
@@ -63,6 +65,24 @@ class IssueDao(id: EntityID<Long>) : LongEntity(id) {
                 severity = issue.severity
                 affectedPath = issue.affectedPath
             }
+
+        /**
+         * Execute the given [query] and return the result as a list of [Issue] objects. This function allows to
+         * load issues from different sources, e.g. the issues from an ORT run or from a scan result. The query must
+         * return the properties of an issue in the following order:
+         * 1. timestamp
+         * 2. source
+         * 3. message
+         * 4. severity
+         * 5. affectedPath (nullable)
+         * 6. identifier type (nullable)
+         * 7. identifier name (nullable)
+         * 8. identifier namespace (nullable)
+         * 9. identifier version (nullable)
+         * 10. worker (nullable)
+         */
+        fun createFromQuery(query: Query): List<Issue> =
+            query.map(ResultRow::toIssue)
     }
 
     var source by IssuesTable.issueSource
@@ -81,5 +101,38 @@ class IssueDao(id: EntityID<Long>) : LongEntity(id) {
         affectedPath = affectedPath,
         identifier = identifier,
         worker = worker
+    )
+}
+
+/**
+ * Create an [Issue] from the fields of this [ResultRow]. Using this function, issues can be created from query
+ * results that may include other data as well. The fields of the result define the issue; they must appear in the
+ * order as described in [IssueDao.createFromQuery].
+ */
+@Suppress("ComplexCondition")
+fun ResultRow.toIssue(): Issue {
+    // The exposed library seems not to fully support fields on alias queries when unions are used.
+    // Use the field indexes in order to extract the values from the ResultRow instead of the field names.
+    // Disadvantage: More fragility, losing type safety at compile time.
+    val columns = fieldIndex.keys.toList()
+
+    val type = this[columns[5]] as String?
+    val name = this[columns[6]] as String?
+    val namespace = this[columns[7]] as String?
+    val version = this[columns[8]] as String?
+    val identifier = if (type == null || name == null || namespace == null || version == null) {
+        null
+    } else {
+        Identifier(type, namespace, name, version)
+    }
+
+    return Issue(
+        timestamp = this[columns[0]] as Instant,
+        source = this[columns[1]] as String,
+        message = this[columns[2]] as String,
+        severity = this[columns[3]] as Severity,
+        affectedPath = this[columns[4]] as String?,
+        worker = this[columns[9]] as String?,
+        identifier = identifier
     )
 }
